Move definition of EventsTablesUnion into separate file
