+++ flowmachine/flowmachine/features/subscriber/subscriber_call_durations.py
@@ -14,7 +14,7 @@ from typing import List
 
 from ...core import JoinToLocation
 from ...utils.utils import get_columns_for_level
-from ..utilities.sets import EventsTablesUnion
+from ..utilities import EventsTablesUnion
 from .metaclasses import SubscriberFeature
 
 valid_stats = {"count", "sum", "avg", "max", "min", "median", "stddev", "variance"}
