+++ flowmachine/flowmachine/features/subscriber/total_subscriber_events.py
@@ -10,7 +10,7 @@ have done over a certain time period.
 
 """
 import flowmachine
-from ..utilities.sets import EventsTablesUnion
+from ..utilities import EventsTablesUnion
 from .metaclasses import SubscriberFeature
