+++ flowmachine/flowmachine/features/utilities/__init__.py
@@ -10,5 +10,6 @@ from .subscriber_locations import subscriber_locations
 from .feature_collection import FeatureCollection
 
 
-from .sets import UniqueSubscribers, EventsTablesUnion, SubscriberLocationSubset
+from .sets import UniqueSubscribers, SubscriberLocationSubset
 from .event_table_subset import EventTableSubset
+from .events_tables_union import EventsTablesUnion
